Classification. Choose tutors
Algorithms for analyze data. Choose proper tutors for math exam
The evaluation metric is ROC AUC
In this competition your task will be to predict the probability for a tutor to be a proper one for preparing for the math exam. You will be given two datasets: train.csv (contains all features and the target) and test.csv (only features).